🌿 [codebase-cleanup] client: publish a module_core testing library and relocate its tests [step 13/45]
