+++ client/app/test/features/session_diffs/widgets/diff_hunk_widget_test.dart
@@ -1,6 +1,6 @@
 import "package:flutter_test/flutter_test.dart";
 import "package:material_ui/material_ui.dart";
-import "package:sesori_dart_core/src/utils/diff/diff_engine.dart";
+import "package:sesori_dart_core/sesori_dart_core.dart";
 
 import "package:sesori_mobile/features/session_diffs/models/diff_file_view_model.dart";
 import "package:sesori_mobile/features/session_diffs/widgets/diff_hunk_widget.dart";
