+++ client/module_core/lib/sesori_dart_core.dart
@@ -36,9 +36,7 @@ export "src/api/bridge_settings_api.dart";
 export "src/api/client/relay_http_client.dart";
 export "src/api/filesystem_api.dart";
 export "src/api/message_image_api.dart";
-export "src/api/notification_api.dart";
 export "src/api/notification_preferences_api.dart";
-export "src/api/plugin_preference_api.dart";
 export "src/api/product_analytics_preference_api.dart";
 export "src/api/project_api.dart";
 export "src/api/session_api.dart";
@@ -47,11 +45,9 @@ export "src/api/storage/notification_preferences_device_id_storage.dart";
 export "src/api/storage/product_analytics_preference_storage.dart";
 export "src/capabilities/notifications/register_token_request.dart";
 export "src/capabilities/relay/relay_client.dart";
-export "src/capabilities/relay/relay_config.dart";
 export "src/capabilities/relay/room_key_storage.dart";
 export "src/capabilities/server_connection/connection_service.dart";
 export "src/capabilities/server_connection/models/connection_status.dart";
-export "src/capabilities/server_connection/models/sse_event.dart";
 export "src/capabilities/server_connection/server_connection_config.dart";
 export "src/capabilities/session/session_service.dart";
 // Capabilities
@@ -114,7 +110,6 @@ export "src/foundation/models/product_analytics/analytics_runtime_capability.dar
 export "src/foundation/models/product_analytics/installation_analytics_event.dart";
 export "src/foundation/models/product_analytics/product_analytics_event.dart";
 export "src/foundation/models/product_analytics/product_analytics_preference.dart";
-export "src/foundation/models/session_options/session_options_request_mode.dart";
 export "src/foundation/platform/analytics_client.dart";
 export "src/foundation/platform/attachment_thumbnail_storage.dart";
 export "src/foundation/platform/image_clipboard.dart";
@@ -135,23 +130,17 @@ export "src/platform/route_source.dart";
 export "src/platform/url_launcher.dart";
 export "src/repositories/analytics_repository.dart";
 export "src/repositories/appearance_store.dart";
-export "src/repositories/bridge_repository.dart";
 export "src/repositories/bridge_settings_repository.dart";
 export "src/repositories/chat_input_mode_store.dart";
 export "src/repositories/composer_draft_repository.dart";
 export "src/repositories/legal_repository.dart";
 export "src/repositories/message_image_repository.dart";
-export "src/repositories/models/analytics_delivery_result.dart";
 export "src/repositories/models/bridge_settings_result.dart";
-export "src/repositories/models/plugin_discovery_snapshot.dart";
 export "src/repositories/models/plugin_management_result.dart";
-export "src/repositories/models/product_analytics_preference_models.dart";
 export "src/repositories/models/repo_provider.dart";
-export "src/repositories/models/session_options_repository_result.dart";
 export "src/repositories/notification_preferences_repository.dart";
 export "src/repositories/notification_repository.dart";
 export "src/repositories/permission_repository.dart";
-export "src/repositories/plugin_preference_repository.dart";
 export "src/repositories/plugin_repository.dart";
 export "src/repositories/product_analytics_preference_repository.dart";
 export "src/repositories/project_repository.dart";
@@ -172,7 +161,6 @@ export "src/services/models/new_session_options_source.dart";
 export "src/services/models/new_session_selection_intent.dart";
 export "src/services/models/product_analytics_state.dart";
 export "src/services/models/session_activity_info.dart";
-export "src/services/models/session_list_item_state.dart";
 export "src/services/new_session_options_service.dart";
 export "src/services/new_session_plugin_service.dart";
 export "src/services/new_session_selection_tracker.dart";
