+++ client/module_core/lib/testing.dart
@@ -1,0 +1,1 @@
+export "src/testing/test_helpers.dart";
