+++ client/module_core/pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   injectable: ^3.0.0
   json_annotation: ^4.12.0
   meta: ^1.19.0
+  mocktail: ^1.0.5
   path: any
   rxdart: ^0.28.0
   sesori_auth:
@@ -34,6 +35,5 @@ dev_dependencies:
   freezed: 4.0.0-dev.3
   injectable_generator: ^3.1.3
   json_serializable: ^6.14.1
-  mocktail: ^1.0.5
   test: ^1.31.1
   fake_async: ^1.3.3
