+++ client/module_core/test/api/client/relay_http_client_test.dart
@@ -1,16 +1,15 @@
 import "dart:async";
 
-import "package:flutter_test/flutter_test.dart";
 import "package:mocktail/mocktail.dart";
 import "package:sesori_auth/sesori_auth.dart";
 import "package:sesori_dart_core/src/api/client/relay_http_client.dart";
+import "package:sesori_dart_core/testing.dart";
 import "package:sesori_shared/sesori_shared.dart";
-
-import "../../../helpers/test_helpers.dart";
+import "package:test/test.dart";
 
 void main() {
   setUpAll(() {
-    registerAllFallbackValues();
+    registerCoreFallbackValues();
     registerFallbackValue(
       const RelayRequest(id: "fake-id", method: "GET", path: "/", headers: {}, body: null),
     );
