+++ client/module_core/test/api/notification_api_test.dart
@@ -1,6 +1,7 @@
 import "package:mocktail/mocktail.dart";
 import "package:sesori_auth/sesori_auth.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/api/notification_api.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
