+++ client/module_core/test/capabilities/relay/relay_client_test.dart
@@ -1,8 +1,7 @@
-import "package:flutter_test/flutter_test.dart";
 import "package:sesori_dart_core/src/capabilities/relay/relay_client.dart";
+import "package:sesori_dart_core/testing.dart";
 import "package:sesori_shared/sesori_shared.dart";
-
-import "../../helpers/test_helpers.dart";
+import "package:test/test.dart";
 
 void main() {
   group("RelayCloseCodes constants", () {
