+++ client/module_core/test/capabilities/relay/room_key_storage_test.dart
@@ -1,17 +1,16 @@
 import "dart:convert";
 import "dart:typed_data";
 
-import "package:flutter_test/flutter_test.dart";
 import "package:mocktail/mocktail.dart";
 import "package:sesori_dart_core/src/capabilities/relay/room_key_storage.dart";
-
-import "../../helpers/test_helpers.dart";
+import "package:sesori_dart_core/testing.dart";
+import "package:test/test.dart";
 
 void main() {
   late MockSecureStorage mockStorage;
   late RoomKeyStorage roomKeyStorage;
 
-  setUpAll(registerAllFallbackValues);
+  setUpAll(registerCoreFallbackValues);
 
   setUp(() {
     mockStorage = MockSecureStorage();
