+++ client/module_core/test/capabilities/server_connection/models/sse_event_test.dart
@@ -1,4 +1,4 @@
-import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/capabilities/server_connection/models/sse_event.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
