+++ client/module_core/test/capabilities/server_connection/sse_event_test.dart
@@ -1,6 +1,6 @@
-import "package:flutter_test/flutter_test.dart";
 import "package:sesori_dart_core/src/capabilities/server_connection/models/sse_event.dart";
 import "package:sesori_shared/sesori_shared.dart";
+import "package:test/test.dart";
 
 void main() {
   group("SseEvent.sessionId lazy getter", () {
