+++ client/module_core/test/repositories/notification_repository_test.dart
@@ -1,5 +1,6 @@
 import "package:mocktail/mocktail.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/api/notification_api.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
