+++ client/module_core/test/repositories/product_analytics_preference_repository_test.dart
@@ -1,6 +1,7 @@
 import "dart:collection";
 
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/repositories/models/product_analytics_preference_models.dart";
 import "package:test/test.dart";
 
 const _userKey = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
