+++ client/module_core/test/routing/analytics_route_listener_test.dart
@@ -4,6 +4,7 @@ import "dart:collection";
 import "package:mocktail/mocktail.dart";
 import "package:rxdart/rxdart.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/repositories/models/analytics_delivery_result.dart";
 import "package:test/test.dart";
 
 class _FakeRouteSource({required AppRouteDef? initialRoute}) implements RouteSource {
