+++ client/module_core/test/services/product_analytics_service_test.dart
@@ -6,6 +6,8 @@ import "package:mocktail/mocktail.dart";
 import "package:rxdart/rxdart.dart";
 import "package:sesori_auth/sesori_auth.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/repositories/models/analytics_delivery_result.dart";
+import "package:sesori_dart_core/src/repositories/models/product_analytics_preference_models.dart";
 import "package:sesori_dart_core/src/services/product_analytics_preference_service.dart";
 import "package:test/test.dart";
