+++ client/module_core/test/services/sse_event_tracker_test.dart
@@ -3,6 +3,7 @@ import "dart:async";
 import "package:mocktail/mocktail.dart";
 import "package:rxdart/rxdart.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
+import "package:sesori_dart_core/src/capabilities/server_connection/models/sse_event.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
 import "../helpers/test_helpers.dart";
